+++ src/utils.js
@@ -21,6 +21,12 @@ export const getArrayRandomLength = (min, max, array) => array.slice(0, (getRand
 
 export const getBooleanValue = () => Boolean(getRandomInteger(0, 1));
 
+export const getStringFromArray = (array, sign) => array.join(`${sign} `).toString();
+
+export const getСapitalizedString = (str) => str.replace(/(^|\s)\S/g, (a) =>a.toUpperCase());
+
+export const getPictureUrl = (dir, picture) => `./images/${dir}/${picture}`;
+
 export const getRuntimeInHours = (runtime) => {
   const hours = Math.floor(runtime / ONE_HOUR);
   const minutes = runtime % ONE_HOUR;
@@ -28,6 +34,20 @@ export const getRuntimeInHours = (runtime) => {
   return `${hours}h ${minutes}m`;
 };
 
-export const getStringFromArray = (array, sign) => array.join(`${sign} `).toString();
+export const getReleaseDate = (date) => {
+  const year = date.getFullYear();
+  const month = date.toLocaleString(`en-US`, {month: `long`});
+  const day = date.getDay();
 
-export const getPictureUrl = (dir, picture) => `./images/${dir}/${picture}`;
+  return `${day} ${month} ${year}`;
+};
+
+export const getCommentDate = (date) => {
+  const year = date.getFullYear();
+  const month = date.getMonth();
+  const day = date.getDay();
+
+  const time = date.toLocaleTimeString([], {hour: `2-digit`, minute: `2-digit`}).replace(`PM`, ``);
+
+  return `${year}/${day}/${month} ${time}`;
+};
